modified:   lib/components/graph_bar/bar_data.dart 	modified:   lib/components/graph_bar/graphbar.dart 	modified:   lib/components/graph_bar/indiv_bar.dart 	modified:   lib/historypage.dart 	modified:   lib/homepage.dart 	modified:   lib/main.dart 	new file:   lib/mongodb_config/firebase_db.dart 	modified:   lib/productmenupage.dart 	modified:   lib/serianumpage.dart 	modified:   lib/settingspage.dart 	modified:   lib/test.dart    modified:   pubspec.lock 	modified:   pubspec.yaml

+++ lib/components/graph_bar/bar_data.dart
@@ -1,13 +1,13 @@
 import 'package:waterdetection/components/graph_bar/indiv_bar.dart';
 
 class BarData {
-  final double MonWater;
-  final double TueWater;
-  final double WedWater;
-  final double ThuWater;
-  final double FriWater;
-  final double SatWater;
-  final double SunWater;
+  final int MonWater;
+  final int TueWater;
+  final int WedWater;
+  final int ThuWater;
+  final int FriWater;
+  final int SatWater;
+  final int SunWater;
   BarData(
       {required this.MonWater,
       required this.TueWater,
