+++ lib/components/graph_bar/indiv_bar.dart
@@ -1,5 +1,5 @@
 class Indiv_Bar {
-  final double y;
+  final int y;
   final int x;
   Indiv_Bar({required this.x, required this.y});
 }
